[TASK] Make all TCA fields of pages excludable
All fields can now be excluded with user permissions. It is possible now
to exclude the fields that promotes our premium version. Please only
exclude those fields, when you are also excluding the other fields. We can
only continue the work on this plugin when we sell premium versions as
well.

Resolves #291